Refactoring and basic unions
